+++ src/main/java/com/settler/Foo.java
@@ -1,7 +1,6 @@
 package com.settler;
 
-/**
- * Created by pivotal on 26/01/2016.
- */
 public interface Foo {
+
+    void bar();
 }