+++ src/main/java/com/settler/FooImpl.java
@@ -2,7 +2,7 @@ package com.settler;
 
 import android.util.Log;
 
-public class FooImpl {
+public class FooImpl implements Foo {
 
     public void bar(){
         Log.d("dadas", "Foobar!!!!!!!!!!±!");