+++ src/main/java/com/settler/PropertyListReceiver.java
@@ -1,0 +1,9 @@
+package com.settler;
+
+import com.settler.api.Property;
+
+public interface PropertyListReceiver {
+
+    void updatePropertyList(Property[] properties);
+
+}