+++ src/main/java/com/settler/SettlerApplicationModule.java
@@ -1,4 +1,15 @@
 package com.settler;
 
-public class SettlerApplicationModule {
+import com.google.inject.Binder;
+import com.google.inject.Module;
+import com.settler.api.client.APIClient;
+import com.settler.api.client.DefaultAPIClient;
+
+public class SettlerApplicationModule implements Module {
+
+    @Override
+    public void configure(Binder binder) {
+        binder.bind(APIClient.class).to(DefaultAPIClient.class);
+        binder.bind(Foo.class).to(FooImpl.class);
+    }
 }