+++ src/test/java/com/settler/SettlerApiTest.java
@@ -1,14 +1,12 @@
-package com.example.activity;
+package com.settler;
 
 import org.junit.Test;
 
-/**
- * Created by pivotal on 25/01/2016.
- */
 public class SettlerApiTest {
 
     @Test
     public void testListAll() {
+//        SettlerService service = new SettlerService();
 
     }