+++ HISTORY.md
@@ -1,2 +1,5 @@
 1. Cloned the starter project from repository https://github.com/robolectric/deckard
 1. Added `useLibrary 'org.apache.http.legacy'` to the `android` section in `build.gradle`
+1. Added JUnit 4 and RoboGuice dependencies
+1. Created tests and a APIClient / APIServive based on Retrofit
+1. Created tests and a SettlerApiService (RoboIntentService) to pass the tests